* qt4/candwin/qt4.cpp   - (CandidateWindow::setIndex): Remove unneeded cast.

+++ qt4/candwin/qt4.cpp
@@ -554,7 +554,7 @@ void CandidateWindow::setIndex( int index )
     // set page
     int newpage = 0;
     if ( displayLimit )
-        newpage = ( int ) candidateIndex / displayLimit;
+        newpage = candidateIndex / displayLimit;
     if ( pageIndex != newpage )
         setPage( newpage );
